requirements.txt
tensorflow-gpu==1.4.0
numpy
nltk
pyyaml